refactor virus scan utility method to saperate module so that it can be re-used
