+++ app/helpers/measure_helper.rb
@@ -10,36 +10,6 @@ module MeasureHelper
     end
   end
 
-  class VirusScannerError < SharedError
-    def initialize
-      front_end_version = {
-        title: "Error Loading Measure",
-        summary: "The measure could not be loaded.",
-        body: "Error: V101. Bonnie has encountered an error while trying to load the measure."
-      }
-      back_end_version = {
-        json: {status: "error", messages: "Cannot perform virus scanning."},
-        status: :bad_request
-      }
-      super(front_end_version: front_end_version, back_end_version: back_end_version, operator_error: true)
-    end
-  end
-
-  class VirusFoundError < SharedError
-    def initialize
-      front_end_version = {
-        title: "Error Loading Measure",
-        summary: "The uploaded file is not a valid Measure Authoring Tool (MAT) export of a FHIR Based Measure.",
-        body: "Error: V100. Please re-package and re-export your FHIR based measure from the MAT and try again."
-      }
-      back_end_version = {
-        json: {status: "error", messages: "Potential virus found in file"},
-        status: :bad_request
-      }
-      super(front_end_version: front_end_version, back_end_version: back_end_version, operator_error: true)
-    end
-  end
-
   class MeasureUpdateMeasureNotFound < SharedError
     def initialize
       front_end_version = {
